+++ ostap/fitting/tests/test_fitting_jackknife.py
@@ -57,23 +57,31 @@ def test_jackknife ( ) :
 
     N = 400 
     dataset = model.generate ( N , sample = False )
-
-    ## prefit the whole dataset
+    
+    more_vars   = { 'vm' : lambda  r, _ : ( r.mean_G - 0.4 ) / 0.1       ,
+                    'vs' : lambda  r, _ :   r.sigma_G        / 0.1 - 1   ,
+                    'vr' : lambda  r, _ :   r.sigma_G * 1    / r.mean_G  } 
+    
+     ## prefit the whole dataset
     res , f = model.fitTo ( dataset , draw = True , nbins = 100 , silent = True , refit = 5 )
 
     ## start Jackknife process 
     results , stats = Toys.make_jackknife  (
-        pdf         = model    ,
-        data        = dataset  , 
+        pdf         = model      ,
+        data        = dataset    , 
         fit_config  = { 'silent' : True , 'refit'   : 5   } ,
         fit_pars    = { 'mean_G' : 0.4  , 'sigma_G' : 0.1 } ,
+        more_vars   = more_vars  , 
         silent      = True , 
         progress    = True )
 
     ## fit the whole sample 
     res , f = model.fitTo ( dataset , draw = True  , nbins = 100 , silent = True , refit = 5 ) 
     ## print the final table
-    Toys.print_jackknife ( res , stats , logger )
+    Toys.print_jackknife ( res   ,
+                           stats ,
+                           morevars = dict ( ( k , more_vars [ k ]( res , model ) ) for k in more_vars ) ,
+                           logger   = logger )
 
     time.sleep ( 2 )